+++ R/CompareQuartets.R
@@ -112,19 +112,14 @@ QuartetState <- function (tips, bips, splits = bips, asRaw = FALSE) {
 #' @return `QuartetStates()` returns a vector listing the status of each 
 #' quartet of leaves (in the order listed by [`AllQuartets()`]) in turn.
 #' @export
-QuartetStates <- function (splits, asRaw = FALSE) {
+QuartetStates <- function (splits, asRaw = FALSE) UseMethod('QuartetStates')
+
+#' @export
+QuartetStates.phylo <- function (splits, asRaw = FALSE) {
   splits <- as.Splits(splits)
-  nTip <- NTip(splits)[1]
-  allQuartets <- AllQuartets(nTip)
+  nTip <- NTip(splits)
   
-  if (is.list(splits)) {
-    nQuartets <- ncol(allQuartets)
-    return(t(vapply(splits, QuartetStates, 
-                  if (asRaw) raw(nQuartets) else integer(nQuartets),
-                  asRaw = asRaw)))
-  }
-  
-  ret <- apply(allQuartets, 2, .Subsplit4, unname(splits), nTip)
+  ret <- quartet_states(splits, nTip)
   
   # Return:
   if (asRaw) {
@@ -134,56 +129,25 @@ QuartetStates <- function (splits, asRaw = FALSE) {
   }
 }
 
-#' Closest to tip 1 in each split
-#' 
-#' @param tips Vector of length four specifying index of tips to consider.
-#' @param splits Splits object.
-#' @param nTip Integer specifying number of splits in `splits`.
-#' @return Raw vector specifying the closest relative of `tips[1]` in each 
-#' split
-#' @importFrom TreeTools NTip
-#' @keywords internal
-.Subsplit4 <- function (tips, splits, nTip = NTip(splits)[1]) {
+#' @rdname QuartetState
+#' @export
+QuartetStates.Splits <- QuartetStates.phylo
 
-  blankMask <- raw((nTip - 1L) %/% 8L + 1L)
-  masks <- as.raw(c(1, 2, 4, 8, 16, 32, 64, 128))
-  tipMask <- vapply(tips, function (tip) {
-    mask <- blankMask
-    element <- (tip - 1L) %/% 8L + 1L
-    mask[element] <- masks[(tip - 1L) %% 8L + 1L]
-    mask
-  }, blankMask)
-  if (is.null(dim(tipMask))) tipMask <- matrix(tipMask, 1L)
+#' @rdname QuartetState
+#' @export
+QuartetStates.list <- function (splits, asRaw = FALSE) {
+  splits <- as.Splits(splits)
+  nTip <- NTip(splits)[1]
   
-  mask12 <- tipMask[, 1] | tipMask[, 2]
-  mask13 <- tipMask[, 1] | tipMask[, 3]
-  mask14 <- tipMask[, 1] | tipMask[, 4]
-  mask23 <- tipMask[, 2] | tipMask[, 3]
-  mask24 <- tipMask[, 2] | tipMask[, 4]
-  mask34 <- tipMask[, 3] | tipMask[, 4]
-  mask <- mask12 | mask34
-  
-  subSplits <- splits & mask
-  
-  ret <- as.raw(0L)
-  for (i in seq_len(nrow(subSplits))) {
-    # Up to twice as fast if we don't remove duplicates
-    split <- subSplits[i, ]
-    if (identical(split, mask12) || identical(split, mask34)) {
-      ret <- as.raw(2L)
-      break
-    } else if (identical(split, mask13) || identical(split, mask24))  {
-      ret <- as.raw(3L)
-      break
-    } else if (identical(split, mask14) || identical(split, mask23))  {
-      ret <- as.raw(4L)
-      break
-    }
-  }
-  
-  # Return:
-  ret
+  nQuartets <- choose(nTip, 4)
+  return(t(vapply(splits, QuartetStates, 
+                  if (asRaw) raw(nQuartets) else integer(nQuartets),
+                  asRaw = asRaw)))
 }
+
+#' @rdname QuartetState
+#' @export
+QuartetStates.multiPhylo <- QuartetStates.list
 
 
 #' Compare quartet states by explicit enumeration